// src/app/modules/transportation/routing/model/route-points-view.ts
/**
 * Transport
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Gets or Sets RoutePointsView
 */
export interface RoutePointsView { 
    /**
     * Gets Or Sets
     */
    name?: string;
    /**
     * Gets Or Sets
     */
    latitude?: number;
    /**
     * Gets Or Sets
     */
    longitude?: number;
    /**
     * Gets Or Sets
     */
    position?: number;
    /**
     * Gets Or Sets
     */
    pointAddress?: string;
}